Decouple image upload from Spree
