+++ app/views/api/product_images/update_product_image.v1.rabl
